+++ src/pages/history/_layouts/index.ts
@@ -1,3 +1,3 @@
-export { default as FirstSection } from './FirstSection';
-export { default as SecondSection } from './SecondSection';
-export { default as ThirdSection } from './ThirdSection';
+export { default as FirstSection } from './historyPage/FirstSection';
+export { default as SecondSection } from './historyPage/SecondSection';
+export { default as ThirdSection } from './historyPage/ThirdSection';
